refactor for easy maintenance
